pyproject.toml
[tool.poetry]
name = "gooddelo_test"
version = "0.1.0"
description = ""
authors = ["Andrei Arnautov <aleksejtalitej@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
python-dotenv = "^1.0.0"
SQLAlchemy-Utils = "^0.41.1"
psycopg2 = "^2.9.6"
uvicorn = "^0.22.0"
SQLAlchemy = "1.4.42"
databases = "^0.7.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
